Splits terms
