+++ packages/server/src/sync/index.js
@@ -4,9 +4,16 @@ const { runListeners } = require('./schedule');
 const events = require('./events');
 const emitter = require('./emitter');
 
+function run() {
+  runJobs();
+  runListeners();
+}
+
 module.exports = {
+  run,
   runJobs,
   runListeners,
+
   events,
   emitter,
 };
